+++ ff4j-store-aws-dynamodb/src/test/java/org/ff4j/dynamodb/FeatureStoreDynamoDBIT.java
@@ -26,7 +26,7 @@ import com.amazonaws.client.builder.AwsClientBuilder;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import org.ff4j.core.FeatureStore;
-import org.ff4j.test.propertystore.PropertyStoreTestSupport;
+import org.ff4j.dynamodb.feature.FeatureStoreDynamoDB;
 import org.ff4j.test.store.FeatureStoreTestSupport;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
